added utils files for each provider, consolidated gemini utils, added dynamic verbosity and reasoning fetcher
