+++ apps/sim/providers/anthropic/index.ts
@@ -2,34 +2,22 @@ import Anthropic from '@anthropic-ai/sdk'
 import { createLogger } from '@/lib/logs/console/logger'
 import type { StreamingExecution } from '@/executor/types'
 import { MAX_TOOL_ITERATIONS } from '@/providers'
+import {
+  checkForForcedToolUsage,
+  createReadableStreamFromAnthropicStream,
+  generateToolUseId,
+} from '@/providers/anthropic/utils'
+import { getProviderDefaultModel, getProviderModels } from '@/providers/models'
+import type {
+  ProviderConfig,
+  ProviderRequest,
+  ProviderResponse,
+  TimeSegment,
+} from '@/providers/types'
+import { prepareToolExecution, prepareToolsWithUsageControl } from '@/providers/utils'
 import { executeTool } from '@/tools'
-import { getProviderDefaultModel, getProviderModels } from '../models'
-import type { ProviderConfig, ProviderRequest, ProviderResponse, TimeSegment } from '../types'
-import { prepareToolExecution, prepareToolsWithUsageControl, trackForcedToolUsage } from '../utils'
 
 const logger = createLogger('AnthropicProvider')
-
-/**
- * Helper to wrap Anthropic streaming into a browser-friendly ReadableStream
- */
-function createReadableStreamFromAnthropicStream(
-  anthropicStream: AsyncIterable<any>
-): ReadableStream {
-  return new ReadableStream({
-    async start(controller) {
-      try {
-        for await (const event of anthropicStream) {
-          if (event.type === 'content_block_delta' && event.delta?.text) {
-            controller.enqueue(new TextEncoder().encode(event.delta.text))
-          }
-        }
-        controller.close()
-      } catch (err) {
-        controller.error(err)
-      }
-    },
-  })
-}
 
 export const anthropicProvider: ProviderConfig = {
   id: 'anthropic',
@@ -47,11 +35,6 @@ export const anthropicProvider: ProviderConfig = {
     }
 
     const anthropic = new Anthropic({ apiKey: request.apiKey })
-
-    // Helper function to generate a simple unique ID for tool uses
-    const generateToolUseId = (toolName: string) => {
-      return `${toolName}-${Date.now()}-${Math.random().toString(36).substring(2, 7)}`
-    }
 
     // Transform messages to Anthropic format
     const messages: any[] = []
@@ -393,44 +376,17 @@ ${fieldDescriptions}
           },
         ]
 
-        // Helper function to check for forced tool usage in Anthropic responses
-        const checkForForcedToolUsage = (response: any, toolChoice: any) => {
-          if (
-            typeof toolChoice === 'object' &&
-            toolChoice !== null &&
-            Array.isArray(response.content)
-          ) {
-            const toolUses = response.content.filter((item: any) => item.type === 'tool_use')
-
-            if (toolUses.length > 0) {
-              // Convert Anthropic tool_use format to a format trackForcedToolUsage can understand
-              const adaptedToolCalls = toolUses.map((tool: any) => ({
-                name: tool.name,
-              }))
-
-              // Convert Anthropic tool_choice format to match OpenAI format for tracking
-              const adaptedToolChoice =
-                toolChoice.type === 'tool' ? { function: { name: toolChoice.name } } : toolChoice
-
-              const result = trackForcedToolUsage(
-                adaptedToolCalls,
-                adaptedToolChoice,
-                logger,
-                'anthropic',
-                forcedTools,
-                usedForcedTools
-              )
-              // Make the behavior consistent with the initial check
-              hasUsedForcedTool = result.hasUsedForcedTool
-              usedForcedTools = result.usedForcedTools
-              return result
-            }
-          }
-          return null
-        }
-
         // Check if a forced tool was used in the first response
-        checkForForcedToolUsage(currentResponse, originalToolChoice)
+        const firstCheckResult = checkForForcedToolUsage(
+          currentResponse,
+          originalToolChoice,
+          forcedTools,
+          usedForcedTools
+        )
+        if (firstCheckResult) {
+          hasUsedForcedTool = firstCheckResult.hasUsedForcedTool
+          usedForcedTools = firstCheckResult.usedForcedTools
+        }
 
         try {
           while (iterationCount < MAX_TOOL_ITERATIONS) {
@@ -576,7 +532,16 @@ ${fieldDescriptions}
             currentResponse = await anthropic.messages.create(nextPayload)
 
             // Check if any forced tools were used in this response
-            checkForForcedToolUsage(currentResponse, nextPayload.tool_choice)
+            const nextCheckResult = checkForForcedToolUsage(
+              currentResponse,
+              nextPayload.tool_choice,
+              forcedTools,
+              usedForcedTools
+            )
+            if (nextCheckResult) {
+              hasUsedForcedTool = nextCheckResult.hasUsedForcedTool
+              usedForcedTools = nextCheckResult.usedForcedTools
+            }
 
             const nextModelEndTime = Date.now()
             const thisModelTime = nextModelEndTime - nextModelStartTime
@@ -746,44 +711,17 @@ ${fieldDescriptions}
         },
       ]
 
-      // Helper function to check for forced tool usage in Anthropic responses
-      const checkForForcedToolUsage = (response: any, toolChoice: any) => {
-        if (
-          typeof toolChoice === 'object' &&
-          toolChoice !== null &&
-          Array.isArray(response.content)
-        ) {
-          const toolUses = response.content.filter((item: any) => item.type === 'tool_use')
-
-          if (toolUses.length > 0) {
-            // Convert Anthropic tool_use format to a format trackForcedToolUsage can understand
-            const adaptedToolCalls = toolUses.map((tool: any) => ({
-              name: tool.name,
-            }))
-
-            // Convert Anthropic tool_choice format to match OpenAI format for tracking
-            const adaptedToolChoice =
-              toolChoice.type === 'tool' ? { function: { name: toolChoice.name } } : toolChoice
-
-            const result = trackForcedToolUsage(
-              adaptedToolCalls,
-              adaptedToolChoice,
-              logger,
-              'anthropic',
-              forcedTools,
-              usedForcedTools
-            )
-            // Make the behavior consistent with the initial check
-            hasUsedForcedTool = result.hasUsedForcedTool
-            usedForcedTools = result.usedForcedTools
-            return result
-          }
-        }
-        return null
-      }
-
       // Check if a forced tool was used in the first response
-      checkForForcedToolUsage(currentResponse, originalToolChoice)
+      const firstCheckResult = checkForForcedToolUsage(
+        currentResponse,
+        originalToolChoice,
+        forcedTools,
+        usedForcedTools
+      )
+      if (firstCheckResult) {
+        hasUsedForcedTool = firstCheckResult.hasUsedForcedTool
+        usedForcedTools = firstCheckResult.usedForcedTools
+      }
 
       try {
         while (iterationCount < MAX_TOOL_ITERATIONS) {
@@ -925,7 +863,16 @@ ${fieldDescriptions}
           currentResponse = await anthropic.messages.create(nextPayload)
 
           // Check if any forced tools were used in this response
-          checkForForcedToolUsage(currentResponse, nextPayload.tool_choice)
+          const nextCheckResult = checkForForcedToolUsage(
+            currentResponse,
+            nextPayload.tool_choice,
+            forcedTools,
+            usedForcedTools
+          )
+          if (nextCheckResult) {
+            hasUsedForcedTool = nextCheckResult.hasUsedForcedTool
+            usedForcedTools = nextCheckResult.usedForcedTools
+          }
 
           const nextModelEndTime = Date.now()
           const thisModelTime = nextModelEndTime - nextModelStartTime
