+++ apps/sim/providers/deepseek/index.ts
@@ -2,6 +2,7 @@ import OpenAI from 'openai'
 import { createLogger } from '@/lib/logs/console/logger'
 import type { StreamingExecution } from '@/executor/types'
 import { MAX_TOOL_ITERATIONS } from '@/providers'
+import { createReadableStreamFromDeepseekStream } from '@/providers/deepseek/utils'
 import { getProviderDefaultModel, getProviderModels } from '@/providers/models'
 import type {
   ProviderConfig,
@@ -17,28 +18,6 @@ import {
 import { executeTool } from '@/tools'
 
 const logger = createLogger('DeepseekProvider')
-
-/**
- * Helper function to convert a DeepSeek (OpenAI-compatible) stream to a ReadableStream
- * of text chunks that can be consumed by the browser.
- */
-function createReadableStreamFromDeepseekStream(deepseekStream: any): ReadableStream {
-  return new ReadableStream({
-    async start(controller) {
-      try {
-        for await (const chunk of deepseekStream) {
-          const content = chunk.choices[0]?.delta?.content || ''
-          if (content) {
-            controller.enqueue(new TextEncoder().encode(content))
-          }
-        }
-        controller.close()
-      } catch (error) {
-        controller.error(error)
-      }
-    },
-  })
-}
 
 export const deepseekProvider: ProviderConfig = {
   id: 'deepseek',
