+++ apps/sim/providers/google/index.ts
@@ -1,6 +1,12 @@
 import { createLogger } from '@/lib/logs/console/logger'
 import type { StreamingExecution } from '@/executor/types'
 import { MAX_TOOL_ITERATIONS } from '@/providers'
+import {
+  cleanSchemaForGemini,
+  convertToGeminiFormat,
+  extractFunctionCall,
+  extractTextContent,
+} from '@/providers/google/utils'
 import { getProviderDefaultModel, getProviderModels } from '@/providers/models'
 import type {
   ProviderConfig,
@@ -718,6 +724,27 @@ export const googleProvider: ProviderConfig = {
                   }
                   logger.info('No function call detected, proceeding with streaming response')
 
+                  // Apply structured output for the final response if responseFormat is specified
+                  // This works regardless of whether tools were forced or auto
+                  if (request.responseFormat) {
+                    streamingPayload.tools = undefined
+                    streamingPayload.toolConfig = undefined
+
+                    const responseFormatSchema =
+                      request.responseFormat.schema || request.responseFormat
+                    const cleanSchema = cleanSchemaForGemini(responseFormatSchema)
+
+                    if (!streamingPayload.generationConfig) {
+                      streamingPayload.generationConfig = {}
+                    }
+                    streamingPayload.generationConfig.responseMimeType = 'application/json'
+                    streamingPayload.generationConfig.responseSchema = cleanSchema
+
+                    logger.info(
+                      'Using structured output for final streaming response after tool execution'
+                    )
+                  }
+
                   const streamingResponse = await fetch(
                     `https://generativelanguage.googleapis.com/v1beta/models/${requestedModel}:streamGenerateContent?key=${request.apiKey}`,
                     {
@@ -899,7 +926,59 @@ export const googleProvider: ProviderConfig = {
                 const nextFunctionCall = extractFunctionCall(nextCandidate)
 
                 if (!nextFunctionCall) {
-                  content = extractTextContent(nextCandidate)
+                  // If responseFormat is specified, make one final request with structured output
+                  if (request.responseFormat) {
+                    const finalPayload = {
+                      ...payload,
+                      contents: nextPayload.contents,
+                      tools: undefined,
+                      toolConfig: undefined,
+                    }
+
+                    const responseFormatSchema =
+                      request.responseFormat.schema || request.responseFormat
+                    const cleanSchema = cleanSchemaForGemini(responseFormatSchema)
+
+                    if (!finalPayload.generationConfig) {
+                      finalPayload.generationConfig = {}
+                    }
+                    finalPayload.generationConfig.responseMimeType = 'application/json'
+                    finalPayload.generationConfig.responseSchema = cleanSchema
+
+                    logger.info('Making final request with structured output after tool execution')
+
+                    const finalResponse = await fetch(
+                      `https://generativelanguage.googleapis.com/v1beta/models/${requestedModel}:generateContent?key=${request.apiKey}`,
+                      {
+                        method: 'POST',
+                        headers: {
+                          'Content-Type': 'application/json',
+                        },
+                        body: JSON.stringify(finalPayload),
+                      }
+                    )
+
+                    if (finalResponse.ok) {
+                      const finalResult = await finalResponse.json()
+                      const finalCandidate = finalResult.candidates?.[0]
+                      content = extractTextContent(finalCandidate)
+
+                      if (finalResult.usageMetadata) {
+                        tokens.prompt += finalResult.usageMetadata.promptTokenCount || 0
+                        tokens.completion += finalResult.usageMetadata.candidatesTokenCount || 0
+                        tokens.total +=
+                          (finalResult.usageMetadata.promptTokenCount || 0) +
+                          (finalResult.usageMetadata.candidatesTokenCount || 0)
+                      }
+                    } else {
+                      logger.warn(
+                        'Failed to get structured output, falling back to regular response'
+                      )
+                      content = extractTextContent(nextCandidate)
+                    }
+                  } else {
+                    content = extractTextContent(nextCandidate)
+                  }
                   break
                 }
 
@@ -985,176 +1064,4 @@ export const googleProvider: ProviderConfig = {
       throw enhancedError
     }
   },
-}
-
-/**
- * Helper function to remove additionalProperties from a schema object
- * and perform a deep copy of the schema to avoid modifying the original
- */
-function cleanSchemaForGemini(schema: any): any {
-  if (schema === null || schema === undefined) return schema
-  if (typeof schema !== 'object') return schema
-  if (Array.isArray(schema)) {
-    return schema.map((item) => cleanSchemaForGemini(item))
-  }
-
-  const cleanedSchema: any = {}
-
-  for (const key in schema) {
-    if (key === 'additionalProperties') continue
-
-    cleanedSchema[key] = cleanSchemaForGemini(schema[key])
-  }
-
-  return cleanedSchema
-}
-
-/**
- * Helper function to extract content from a Gemini response, handling structured output
- */
-function extractTextContent(candidate: any): string {
-  if (!candidate?.content?.parts) return ''
-
-  if (candidate.content.parts?.length === 1 && candidate.content.parts[0].text) {
-    const text = candidate.content.parts[0].text
-    if (text && (text.trim().startsWith('{') || text.trim().startsWith('['))) {
-      try {
-        JSON.parse(text) // Validate JSON
-        return text // Return valid JSON as-is
-      } catch (_e) {
-        /* Not valid JSON, continue with normal extraction */
-      }
-    }
-  }
-
-  return candidate.content.parts
-    .filter((part: any) => part.text)
-    .map((part: any) => part.text)
-    .join('\n')
-}
-
-/**
- * Helper function to extract a function call from a Gemini response
- */
-function extractFunctionCall(candidate: any): { name: string; args: any } | null {
-  if (!candidate?.content?.parts) return null
-
-  for (const part of candidate.content.parts) {
-    if (part.functionCall) {
-      const args = part.functionCall.args || {}
-      if (
-        typeof part.functionCall.args === 'string' &&
-        part.functionCall.args.trim().startsWith('{')
-      ) {
-        try {
-          return { name: part.functionCall.name, args: JSON.parse(part.functionCall.args) }
-        } catch (_e) {
-          return { name: part.functionCall.name, args: part.functionCall.args }
-        }
-      }
-      return { name: part.functionCall.name, args }
-    }
-  }
-
-  if (candidate.content.function_call) {
-    const args =
-      typeof candidate.content.function_call.arguments === 'string'
-        ? JSON.parse(candidate.content.function_call.arguments || '{}')
-        : candidate.content.function_call.arguments || {}
-    return { name: candidate.content.function_call.name, args }
-  }
-
-  return null
-}
-
-/**
- * Convert OpenAI-style request format to Gemini format
- */
-function convertToGeminiFormat(request: ProviderRequest): {
-  contents: any[]
-  tools: any[] | undefined
-  systemInstruction: any | undefined
-} {
-  const contents = []
-  let systemInstruction
-
-  if (request.systemPrompt) {
-    systemInstruction = { parts: [{ text: request.systemPrompt }] }
-  }
-
-  if (request.context) {
-    contents.push({ role: 'user', parts: [{ text: request.context }] })
-  }
-
-  if (request.messages && request.messages.length > 0) {
-    for (const message of request.messages) {
-      if (message.role === 'system') {
-        if (!systemInstruction) {
-          systemInstruction = { parts: [{ text: message.content }] }
-        } else {
-          systemInstruction.parts[0].text = `${systemInstruction.parts[0].text || ''}\n${message.content}`
-        }
-      } else if (message.role === 'user' || message.role === 'assistant') {
-        const geminiRole = message.role === 'user' ? 'user' : 'model'
-
-        if (message.content) {
-          contents.push({ role: geminiRole, parts: [{ text: message.content }] })
-        }
-
-        if (message.role === 'assistant' && message.tool_calls && message.tool_calls.length > 0) {
-          const functionCalls = message.tool_calls.map((toolCall) => ({
-            functionCall: {
-              name: toolCall.function?.name,
-              args: JSON.parse(toolCall.function?.arguments || '{}'),
-            },
-          }))
-
-          contents.push({ role: 'model', parts: functionCalls })
-        }
-      } else if (message.role === 'tool') {
-        contents.push({
-          role: 'user',
-          parts: [{ text: `Function result: ${message.content}` }],
-        })
-      }
-    }
-  }
-
-  const tools = request.tools?.map((tool) => {
-    const toolParameters = { ...(tool.parameters || {}) }
-
-    if (toolParameters.properties) {
-      const properties = { ...toolParameters.properties }
-      const required = toolParameters.required ? [...toolParameters.required] : []
-
-      for (const key in properties) {
-        const prop = properties[key] as any
-
-        if (prop.default !== undefined) {
-          const { default: _, ...cleanProp } = prop
-          properties[key] = cleanProp
-        }
-      }
-
-      const parameters = {
-        type: toolParameters.type || 'object',
-        properties,
-        ...(required.length > 0 ? { required } : {}),
-      }
-
-      return {
-        name: tool.id,
-        description: tool.description || `Execute the ${tool.id} function`,
-        parameters: cleanSchemaForGemini(parameters),
-      }
-    }
-
-    return {
-      name: tool.id,
-      description: tool.description || `Execute the ${tool.id} function`,
-      parameters: cleanSchemaForGemini(toolParameters),
-    }
-  })
-
-  return { contents, tools, systemInstruction }
 }
