+++ apps/sim/providers/mistral/utils.ts
@@ -1,0 +1,39 @@
+/**
+ * Creates a ReadableStream from a Mistral AI streaming response
+ * @param mistralStream - The Mistral AI stream object
+ * @param onComplete - Optional callback when streaming completes
+ * @returns A ReadableStream that yields text chunks
+ */
+export function createReadableStreamFromMistralStream(
+  mistralStream: any,
+  onComplete?: (content: string, usage?: any) => void
+): ReadableStream {
+  let fullContent = ''
+  let usageData: any = null
+
+  return new ReadableStream({
+    async start(controller) {
+      try {
+        for await (const chunk of mistralStream) {
+          if (chunk.usage) {
+            usageData = chunk.usage
+          }
+
+          const content = chunk.choices[0]?.delta?.content || ''
+          if (content) {
+            fullContent += content
+            controller.enqueue(new TextEncoder().encode(content))
+          }
+        }
+
+        if (onComplete) {
+          onComplete(fullContent, usageData)
+        }
+
+        controller.close()
+      } catch (error) {
+        controller.error(error)
+      }
+    },
+  })
+}
