+++ apps/sim/providers/openai/utils.ts
@@ -1,0 +1,37 @@
+/**
+ * Helper function to convert an OpenAI stream to a standard ReadableStream
+ * and collect completion metrics
+ */
+export function createReadableStreamFromOpenAIStream(
+  openaiStream: any,
+  onComplete?: (content: string, usage?: any) => void
+): ReadableStream {
+  let fullContent = ''
+  let usageData: any = null
+
+  return new ReadableStream({
+    async start(controller) {
+      try {
+        for await (const chunk of openaiStream) {
+          if (chunk.usage) {
+            usageData = chunk.usage
+          }
+
+          const content = chunk.choices[0]?.delta?.content || ''
+          if (content) {
+            fullContent += content
+            controller.enqueue(new TextEncoder().encode(content))
+          }
+        }
+
+        if (onComplete) {
+          onComplete(fullContent, usageData)
+        }
+
+        controller.close()
+      } catch (error) {
+        controller.error(error)
+      }
+    },
+  })
+}
