+++ apps/sim/providers/vertex/index.ts
@@ -2,6 +2,12 @@ import { env } from '@/lib/core/config/env'
 import { createLogger } from '@/lib/logs/console/logger'
 import type { StreamingExecution } from '@/executor/types'
 import { MAX_TOOL_ITERATIONS } from '@/providers'
+import {
+  cleanSchemaForGemini,
+  convertToGeminiFormat,
+  extractFunctionCall,
+  extractTextContent,
+} from '@/providers/google/utils'
 import { getProviderDefaultModel, getProviderModels } from '@/providers/models'
 import type {
   ProviderConfig,
@@ -14,238 +20,10 @@ import {
   prepareToolsWithUsageControl,
   trackForcedToolUsage,
 } from '@/providers/utils'
+import { buildVertexEndpoint, createReadableStreamFromVertexStream } from '@/providers/vertex/utils'
 import { executeTool } from '@/tools'
 
 const logger = createLogger('VertexProvider')
-
-/**
- * Creates a ReadableStream from Vertex AI's Gemini stream response
- */
-function createReadableStreamFromVertexStream(
-  response: Response,
-  onComplete?: (
-    content: string,
-    usage?: { promptTokenCount?: number; candidatesTokenCount?: number; totalTokenCount?: number }
-  ) => void
-): ReadableStream<Uint8Array> {
-  const reader = response.body?.getReader()
-  if (!reader) {
-    throw new Error('Failed to get reader from response body')
-  }
-
-  return new ReadableStream({
-    async start(controller) {
-      try {
-        let buffer = ''
-        let fullContent = ''
-        let usageData: {
-          promptTokenCount?: number
-          candidatesTokenCount?: number
-          totalTokenCount?: number
-        } | null = null
-
-        while (true) {
-          const { done, value } = await reader.read()
-          if (done) {
-            if (buffer.trim()) {
-              try {
-                const data = JSON.parse(buffer.trim())
-                if (data.usageMetadata) {
-                  usageData = data.usageMetadata
-                }
-                const candidate = data.candidates?.[0]
-                if (candidate?.content?.parts) {
-                  const functionCall = extractFunctionCall(candidate)
-                  if (functionCall) {
-                    logger.debug(
-                      'Function call detected in final buffer, ending stream to execute tool',
-                      {
-                        functionName: functionCall.name,
-                      }
-                    )
-                    if (onComplete) onComplete(fullContent, usageData || undefined)
-                    controller.close()
-                    return
-                  }
-                  const content = extractTextContent(candidate)
-                  if (content) {
-                    fullContent += content
-                    controller.enqueue(new TextEncoder().encode(content))
-                  }
-                }
-              } catch (e) {
-                if (buffer.trim().startsWith('[')) {
-                  try {
-                    const dataArray = JSON.parse(buffer.trim())
-                    if (Array.isArray(dataArray)) {
-                      for (const item of dataArray) {
-                        if (item.usageMetadata) {
-                          usageData = item.usageMetadata
-                        }
-                        const candidate = item.candidates?.[0]
-                        if (candidate?.content?.parts) {
-                          const functionCall = extractFunctionCall(candidate)
-                          if (functionCall) {
-                            logger.debug(
-                              'Function call detected in array item, ending stream to execute tool',
-                              {
-                                functionName: functionCall.name,
-                              }
-                            )
-                            if (onComplete) onComplete(fullContent, usageData || undefined)
-                            controller.close()
-                            return
-                          }
-                          const content = extractTextContent(candidate)
-                          if (content) {
-                            fullContent += content
-                            controller.enqueue(new TextEncoder().encode(content))
-                          }
-                        }
-                      }
-                    }
-                  } catch (arrayError) {
-                    // Buffer is not valid JSON array
-                  }
-                }
-              }
-            }
-            if (onComplete) onComplete(fullContent, usageData || undefined)
-            controller.close()
-            break
-          }
-
-          const text = new TextDecoder().decode(value)
-          buffer += text
-
-          let searchIndex = 0
-          while (searchIndex < buffer.length) {
-            const openBrace = buffer.indexOf('{', searchIndex)
-            if (openBrace === -1) break
-
-            let braceCount = 0
-            let inString = false
-            let escaped = false
-            let closeBrace = -1
-
-            for (let i = openBrace; i < buffer.length; i++) {
-              const char = buffer[i]
-
-              if (!inString) {
-                if (char === '"' && !escaped) {
-                  inString = true
-                } else if (char === '{') {
-                  braceCount++
-                } else if (char === '}') {
-                  braceCount--
-                  if (braceCount === 0) {
-                    closeBrace = i
-                    break
-                  }
-                }
-              } else {
-                if (char === '"' && !escaped) {
-                  inString = false
-                }
-              }
-
-              escaped = char === '\\' && !escaped
-            }
-
-            if (closeBrace !== -1) {
-              const jsonStr = buffer.substring(openBrace, closeBrace + 1)
-
-              try {
-                const data = JSON.parse(jsonStr)
-
-                if (data.usageMetadata) {
-                  usageData = data.usageMetadata
-                }
-
-                const candidate = data.candidates?.[0]
-
-                if (candidate?.finishReason === 'UNEXPECTED_TOOL_CALL') {
-                  logger.warn(
-                    'Vertex AI returned UNEXPECTED_TOOL_CALL - model attempted to call a tool that was not provided',
-                    {
-                      finishReason: candidate.finishReason,
-                      hasContent: !!candidate?.content,
-                      hasParts: !!candidate?.content?.parts,
-                    }
-                  )
-                  const textContent = extractTextContent(candidate)
-                  if (textContent) {
-                    fullContent += textContent
-                    controller.enqueue(new TextEncoder().encode(textContent))
-                  }
-                  if (onComplete) onComplete(fullContent, usageData || undefined)
-                  controller.close()
-                  return
-                }
-
-                if (candidate?.content?.parts) {
-                  const functionCall = extractFunctionCall(candidate)
-                  if (functionCall) {
-                    logger.debug(
-                      'Function call detected in stream, ending stream to execute tool',
-                      {
-                        functionName: functionCall.name,
-                      }
-                    )
-                    if (onComplete) onComplete(fullContent, usageData || undefined)
-                    controller.close()
-                    return
-                  }
-                  const content = extractTextContent(candidate)
-                  if (content) {
-                    fullContent += content
-                    controller.enqueue(new TextEncoder().encode(content))
-                  }
-                }
-              } catch (e) {
-                logger.error('Error parsing JSON from stream', {
-                  error: e instanceof Error ? e.message : String(e),
-                  jsonPreview: jsonStr.substring(0, 200),
-                })
-              }
-
-              buffer = buffer.substring(closeBrace + 1)
-              searchIndex = 0
-            } else {
-              break
-            }
-          }
-        }
-      } catch (e) {
-        logger.error('Error reading Vertex AI stream', {
-          error: e instanceof Error ? e.message : String(e),
-        })
-        controller.error(e)
-      }
-    },
-    async cancel() {
-      await reader.cancel()
-    },
-  })
-}
-
-/**
- * Build Vertex AI endpoint URL
- */
-function buildVertexEndpoint(
-  project: string,
-  location: string,
-  model: string,
-  isStreaming: boolean
-): string {
-  const action = isStreaming ? 'streamGenerateContent' : 'generateContent'
-
-  if (location === 'global') {
-    return `https://aiplatform.googleapis.com/v1/projects/${project}/locations/global/publishers/google/models/${model}:${action}`
-  }
-
-  return `https://${location}-aiplatform.googleapis.com/v1/projects/${project}/locations/${location}/publishers/google/models/${model}:${action}`
-}
 
 /**
  * Vertex AI provider configuration
@@ -763,6 +541,27 @@ export const vertexProvider: ProviderConfig = {
 
                   logger.info('No function call detected, proceeding with streaming response')
 
+                  // Apply structured output for the final response if responseFormat is specified
+                  // This works regardless of whether tools were forced or auto
+                  if (request.responseFormat) {
+                    streamingPayload.tools = undefined
+                    streamingPayload.toolConfig = undefined
+
+                    const responseFormatSchema =
+                      request.responseFormat.schema || request.responseFormat
+                    const cleanSchema = cleanSchemaForGemini(responseFormatSchema)
+
+                    if (!streamingPayload.generationConfig) {
+                      streamingPayload.generationConfig = {}
+                    }
+                    streamingPayload.generationConfig.responseMimeType = 'application/json'
+                    streamingPayload.generationConfig.responseSchema = cleanSchema
+
+                    logger.info(
+                      'Using structured output for final streaming response after tool execution'
+                    )
+                  }
+
                   const streamEndpoint = buildVertexEndpoint(
                     vertexProject,
                     vertexLocation,
@@ -954,7 +753,64 @@ export const vertexProvider: ProviderConfig = {
                 const nextFunctionCall = extractFunctionCall(nextCandidate)
 
                 if (!nextFunctionCall) {
-                  content = extractTextContent(nextCandidate)
+                  // If responseFormat is specified, make one final request with structured output
+                  if (request.responseFormat) {
+                    const finalPayload = {
+                      ...payload,
+                      contents: nextPayload.contents,
+                      tools: undefined,
+                      toolConfig: undefined,
+                    }
+
+                    const responseFormatSchema =
+                      request.responseFormat.schema || request.responseFormat
+                    const cleanSchema = cleanSchemaForGemini(responseFormatSchema)
+
+                    if (!finalPayload.generationConfig) {
+                      finalPayload.generationConfig = {}
+                    }
+                    finalPayload.generationConfig.responseMimeType = 'application/json'
+                    finalPayload.generationConfig.responseSchema = cleanSchema
+
+                    logger.info('Making final request with structured output after tool execution')
+
+                    const finalEndpoint = buildVertexEndpoint(
+                      vertexProject,
+                      vertexLocation,
+                      requestedModel,
+                      false
+                    )
+
+                    const finalResponse = await fetch(finalEndpoint, {
+                      method: 'POST',
+                      headers: {
+                        'Content-Type': 'application/json',
+                        Authorization: `Bearer ${request.apiKey}`,
+                      },
+                      body: JSON.stringify(finalPayload),
+                    })
+
+                    if (finalResponse.ok) {
+                      const finalResult = await finalResponse.json()
+                      const finalCandidate = finalResult.candidates?.[0]
+                      content = extractTextContent(finalCandidate)
+
+                      if (finalResult.usageMetadata) {
+                        tokens.prompt += finalResult.usageMetadata.promptTokenCount || 0
+                        tokens.completion += finalResult.usageMetadata.candidatesTokenCount || 0
+                        tokens.total +=
+                          (finalResult.usageMetadata.promptTokenCount || 0) +
+                          (finalResult.usageMetadata.candidatesTokenCount || 0)
+                      }
+                    } else {
+                      logger.warn(
+                        'Failed to get structured output, falling back to regular response'
+                      )
+                      content = extractTextContent(nextCandidate)
+                    }
+                  } else {
+                    content = extractTextContent(nextCandidate)
+                  }
                   break
                 }
 
@@ -1040,174 +896,4 @@ export const vertexProvider: ProviderConfig = {
       throw enhancedError
     }
   },
-}
-
-/**
- * Helper function to remove additionalProperties from a schema object
- */
-function cleanSchemaForGemini(schema: any): any {
-  if (schema === null || schema === undefined) return schema
-  if (typeof schema !== 'object') return schema
-  if (Array.isArray(schema)) {
-    return schema.map((item) => cleanSchemaForGemini(item))
-  }
-
-  const cleanedSchema: any = {}
-
-  for (const key in schema) {
-    if (key === 'additionalProperties') continue
-    cleanedSchema[key] = cleanSchemaForGemini(schema[key])
-  }
-
-  return cleanedSchema
-}
-
-/**
- * Helper function to extract content from a Gemini response
- */
-function extractTextContent(candidate: any): string {
-  if (!candidate?.content?.parts) return ''
-
-  if (candidate.content.parts?.length === 1 && candidate.content.parts[0].text) {
-    const text = candidate.content.parts[0].text
-    if (text && (text.trim().startsWith('{') || text.trim().startsWith('['))) {
-      try {
-        JSON.parse(text)
-        return text
-      } catch (_e) {
-        /* Not valid JSON, continue with normal extraction */
-      }
-    }
-  }
-
-  return candidate.content.parts
-    .filter((part: any) => part.text)
-    .map((part: any) => part.text)
-    .join('\n')
-}
-
-/**
- * Helper function to extract a function call from a Gemini response
- */
-function extractFunctionCall(candidate: any): { name: string; args: any } | null {
-  if (!candidate?.content?.parts) return null
-
-  for (const part of candidate.content.parts) {
-    if (part.functionCall) {
-      const args = part.functionCall.args || {}
-      if (
-        typeof part.functionCall.args === 'string' &&
-        part.functionCall.args.trim().startsWith('{')
-      ) {
-        try {
-          return { name: part.functionCall.name, args: JSON.parse(part.functionCall.args) }
-        } catch (_e) {
-          return { name: part.functionCall.name, args: part.functionCall.args }
-        }
-      }
-      return { name: part.functionCall.name, args }
-    }
-  }
-
-  if (candidate.content.function_call) {
-    const args =
-      typeof candidate.content.function_call.arguments === 'string'
-        ? JSON.parse(candidate.content.function_call.arguments || '{}')
-        : candidate.content.function_call.arguments || {}
-    return { name: candidate.content.function_call.name, args }
-  }
-
-  return null
-}
-
-/**
- * Convert OpenAI-style request format to Gemini format
- */
-function convertToGeminiFormat(request: ProviderRequest): {
-  contents: any[]
-  tools: any[] | undefined
-  systemInstruction: any | undefined
-} {
-  const contents = []
-  let systemInstruction
-
-  if (request.systemPrompt) {
-    systemInstruction = { parts: [{ text: request.systemPrompt }] }
-  }
-
-  if (request.context) {
-    contents.push({ role: 'user', parts: [{ text: request.context }] })
-  }
-
-  if (request.messages && request.messages.length > 0) {
-    for (const message of request.messages) {
-      if (message.role === 'system') {
-        if (!systemInstruction) {
-          systemInstruction = { parts: [{ text: message.content }] }
-        } else {
-          systemInstruction.parts[0].text = `${systemInstruction.parts[0].text || ''}\n${message.content}`
-        }
-      } else if (message.role === 'user' || message.role === 'assistant') {
-        const geminiRole = message.role === 'user' ? 'user' : 'model'
-
-        if (message.content) {
-          contents.push({ role: geminiRole, parts: [{ text: message.content }] })
-        }
-
-        if (message.role === 'assistant' && message.tool_calls && message.tool_calls.length > 0) {
-          const functionCalls = message.tool_calls.map((toolCall) => ({
-            functionCall: {
-              name: toolCall.function?.name,
-              args: JSON.parse(toolCall.function?.arguments || '{}'),
-            },
-          }))
-
-          contents.push({ role: 'model', parts: functionCalls })
-        }
-      } else if (message.role === 'tool') {
-        contents.push({
-          role: 'user',
-          parts: [{ text: `Function result: ${message.content}` }],
-        })
-      }
-    }
-  }
-
-  const tools = request.tools?.map((tool) => {
-    const toolParameters = { ...(tool.parameters || {}) }
-
-    if (toolParameters.properties) {
-      const properties = { ...toolParameters.properties }
-      const required = toolParameters.required ? [...toolParameters.required] : []
-
-      for (const key in properties) {
-        const prop = properties[key] as any
-
-        if (prop.default !== undefined) {
-          const { default: _, ...cleanProp } = prop
-          properties[key] = cleanProp
-        }
-      }
-
-      const parameters = {
-        type: toolParameters.type || 'object',
-        properties,
-        ...(required.length > 0 ? { required } : {}),
-      }
-
-      return {
-        name: tool.id,
-        description: tool.description || `Execute the ${tool.id} function`,
-        parameters: cleanSchemaForGemini(parameters),
-      }
-    }
-
-    return {
-      name: tool.id,
-      description: tool.description || `Execute the ${tool.id} function`,
-      parameters: cleanSchemaForGemini(toolParameters),
-    }
-  })
-
-  return { contents, tools, systemInstruction }
 }
